refactor: Home2Feed
