+++ app/[locale]/feed/page.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/views/feedPage/ui/FeedPage'
