--- app/[locale]/home/page.tsx
@@ -1,1 +1,0 @@
-export { default } from '@/views/homePage/ui/HomePage'
