+++ src/shared/assets/icons/FeedIcon.tsx
@@ -1,6 +1,6 @@
 import { Ref, SVGProps, forwardRef, memo } from 'react'
 
-const HomeIcon = (props: SVGProps<SVGSVGElement>, ref: Ref<SVGSVGElement>) => (
+const FeedIcon = (props: SVGProps<SVGSVGElement>, ref: Ref<SVGSVGElement>) => (
   <svg
     {...props}
     fill={'none'}
@@ -27,7 +27,7 @@ const HomeIcon = (props: SVGProps<SVGSVGElement>, ref: Ref<SVGSVGElement>) => (
   </svg>
 )
 
-const ForwardRef = forwardRef(HomeIcon)
+const ForwardRef = forwardRef(FeedIcon)
 const Memo = memo(ForwardRef)
 
 export default Memo
