+++ src/views/feedPage/ui/feedPage.module.scss
